fixed the epipe error and refactored into a cleaner module pattern
